feat: added pull services in spec.yaml
