+++ actions/docker.action.ts
@@ -9,6 +9,7 @@ import * as chalk from 'chalk';
 import * as fs from 'fs';
 import * as yaml from 'js-yaml';
 import { join } from 'path';
+import { getSpecFilePath } from '../lib/utils/specFilePath';
 
 export class DockerAction extends AbstractAction {
   private manager!: AbstractPackageManager;
@@ -17,7 +18,7 @@ export class DockerAction extends AbstractAction {
     
     const serviceName = commandInputs[0].value as string;
 
-    const specFilePath = await this.getSpecFilePath();
+    const specFilePath = await getSpecFilePath();
     
     if (specFilePath) {
       await this.updateSpecFile(specFilePath, serviceName);
@@ -38,20 +39,6 @@ export class DockerAction extends AbstractAction {
       if (error && error.message) {
       console.error(chalk.red(error.message));
       }
-    }
-  }
-  private async getSpecFilePath(): Promise<string | null> {
-    try {
-      const nestCliConfig = await fs.promises.readFile('nest-cli.json', 'utf-8');
-      const config = JSON.parse(nestCliConfig);
-      if (config.specFile) {
-        return config.specFile;
-      } else {
-        return null;
-      }
-    } catch (error) {
-      console.error(chalk.red('Error reading nest-cli.json'), error);
-      return null;
     }
   }
